# Reimplement integer length