https://twitter.com/ChrisGSeaton/status/1398459785092153346

Not as good as I was expecting. Seems to warm up a lot quicker though. What do you think?

```
Warming up --------------------------------------
              length   913.032k i/100ms
Calculating -------------------------------------
              length     51.007M (± 8.1%) i/s -    252.910M in   4.997260s

Warming up --------------------------------------
              length     5.735M i/100ms
Calculating -------------------------------------
              length     57.091M (± 2.8%) i/s -    286.766M in   5.027562s
```