@eregon 2021-05-29
If we can have a node for it, we could have an inline cache specialization (with a fairly small limit like 1, 2 or 3) which guards against the previous and next powers of ten, and just return a constant for the length.

For Time formatting though it's usually fixed-width zero-padded anyway though, so it can probably be simpler:
```
> Time.now.strftime "%m"
=> "05"
```
Probably we'd want a variant of LazyIntRope like ZeroPaddedLazyIntRope or so.