@chrisseaton 2021-05-29
```ruby
require 'benchmark/ips'

puts RUBY_DESCRIPTION

n = 0

Benchmark.ips do |x|
  x.report('length') do
    n = rand(1000000000).to_s.length
  end
end
```