@chrisseaton 2021-05-29
> For Time formatting though it's usually fixed-width zero-padded anyway though, so it can probably be simpler:

The padded numbers in time formatting are in `RopeConstants` so are just normal leaf ropes. The year has a fixed constant length.

https://github.com/oracle/truffleruby/pull/2361/files#diff-2165f1135ce1cc7280599d50b87438aee142f40977cc937dbfd175a8c79795fdR664

> Probably we'd want a variant of LazyIntRope like ZeroPaddedLazyIntRope or so.

Is there any cost to adding more rope types? I can think of lots to add - quickest win could be an n-way concat, for example. Can we add tens of them?