@eregon 2021-05-29
> The padded numbers in time formatting are in RopeConstants so are just normal leaf ropes.

Right, `RopeConstants.paddedNumber` seems great (I did not review that PR in details yet).

> Is there any cost to adding more rope types?
> I can think of lots to add - quickest win could be an n-way concat, for example. Can we add tens of them?

Yes, all nodes operating on specific Ropes must handle the new type.
Also, we are planning to migrate to TruffleString that becomes available (it's making good progress), and that will likely be harder if we have more Rope variants.
There is also of course a risk that some new Rope type isn't handled somewhere and that could cause bugs.
So I think we should only add what's really needed or very beneficial for performance.