@eregon 2021-05-31
The benchmark in https://github.com/oracle/truffleruby/pull/2365#issuecomment-850820680 might be dominated by the cost of `rand`. Could you try something like e.g.
```ruby
require 'benchmark/ips'

puts RUBY_DESCRIPTION

n = 0
numbers = 100.times.map { rand(1000000000) }

Benchmark.ips do |x|
  x.report('length') do
    numbers.each { |v| n = v.to_s.length }
  end
end
```