@chrisseaton 2021-05-31
```
ruby 2.7.3p183 (2021-04-05 revision 6847ee089d) [x86_64-darwin20]
Warming up --------------------------------------
              length     6.195k i/100ms
Calculating -------------------------------------
              length     62.706k (± 3.3%) i/s -    315.945k in   5.045198s

truffleruby 21.2.0-dev-7d950d5b, like ruby 2.7.3, GraalVM CE JVM [x86_64-darwin]
Warming up --------------------------------------
              length   651.096k i/100ms
Calculating -------------------------------------
              length      6.200M (± 6.0%) i/s -     31.253M in   5.062144s

truffleruby 21.2.0-dev-7d950d5b, like ruby 2.7.3, GraalVM CE JVM [x86_64-darwin]
Warming up --------------------------------------
              length   505.515k i/100ms
Calculating -------------------------------------
              length      6.151M (±11.5%) i/s -     30.331M in   5.001263s
```

Not statistically significant. Yeah probably not worth it then. Seems crazy there isn't a better way to do this...