@eregon 2021-06-01
I used a variant of the benchmark to try to measure more precisely:
```ruby
# truffleruby_primitives: true

require 'benchmark/ips'

puts RUBY_DESCRIPTION

small_numbers = [1, 9, 11, 99, 111, 999, 1111, 9999]
rand = Random.new(0)
random_numbers = 10.times.map { rand.rand(1000000000) }
p random_numbers

Benchmark.ips do |x|
  x.iterations = 2
  
  x.report('length small rand') do
    small_numbers.each { |v| Primitive.blackhole v.to_s.length }
  end
  
  x.report('length big rand') do
    random_numbers.each { |v| Primitive.blackhole v.to_s.length }
  end
end
```

Before:
```
   length small rand     68.602M (± 2.2%) i/s -    344.101M in   5.018525s
     length big rand     89.620M (± 0.1%) i/s -    456.744M in   5.096464s
   length small rand     68.804M (± 0.9%) i/s -    344.101M in   5.001580s
     length big rand     89.554M (± 0.1%) i/s -    447.788M in   5.000233s
```

This PR:
```
   length small rand     76.752M (± 0.1%) i/s -    390.447M in   5.087134s
     length big rand     65.355M (± 1.9%) i/s -    326.767M in   5.001952s
   length small rand     76.581M (± 1.3%) i/s -    390.447M in   5.099411s
     length big rand     65.494M (± 0.1%) i/s -    333.302M in   5.089015s
```
So small is slightly faster, big much slower.

Using the improved variant from https://github.com/lemire/Code-used-on-Daniel-Lemire-s-blog/blob/master/2021/05/28/digitcount.java#L7-L13 :
```
   length small rand     97.498M (± 0.3%) i/s -    495.377M in   5.080910s
     length big rand     72.909M (± 0.7%) i/s -    370.870M in   5.086982s
   length small rand     97.402M (± 0.9%) i/s -    495.377M in   5.086326s
     length big rand     72.779M (± 0.5%) i/s -    370.870M in   5.095968s

```
Small is much faster, big is still quite a bit slower.

I guess in practice it's more frequent to have smaller numbers so it might be a good trade-off, but unsure.
The code is obviously less clear.

---

Trying to remove the bound checks with `& 0xf` and making the array bigger doesn't to change the results:
```
   length small rand     97.363M (± 0.3%) i/s -    491.135M in   5.044447s
     length big rand     68.203M (± 4.1%) i/s -    341.794M in   5.020171s
   length small rand     92.409M (± 9.3%) i/s -    456.670M in   4.999764s
     length big rand     68.552M (± 4.3%) i/s -    348.496M in   5.094852s
vs
   length small rand     97.411M (± 0.3%) i/s -    495.123M in   5.082838s
     length big rand     69.195M (± 1.6%) i/s -    349.446M in   5.051373s
   length small rand     97.117M (± 2.0%) i/s -    485.414M in   5.000554s
     length big rand     68.464M (± 2.4%) i/s -    342.315M in   5.002983s
```