@eregon 2021-06-03
Unfortunately this does not work for `Integer.MIN_VALUE = -2147483648`, because `-value` is just the same as `value` and the increment after the lookup table does not work. Probably we need to handle it as a special case, similar to before.