@chrisseaton 2021-06-04
Possible new idea https://lemire.me/blog/2021/06/03/computing-the-number-of-digits-of-an-integer-even-faster/.