@eregon 2021-06-04
Great, that is quite a bit faster! I'll update to that.
```
   length small rand    116.362M (± 0.8%) i/s -    583.549M in   5.015233s
     length big rand     94.310M (± 1.9%) i/s -    473.265M in   5.020145s
   length small rand    116.494M (± 0.6%) i/s -    583.549M in   5.009469s
     length big rand     94.256M (± 2.7%) i/s -    473.265M in   5.025599s
```

So now it's 1.7x faster for small, and 1.05x faster for big than `Before`, and the first time it's faster than Before for big.