@chrisseaton 2021-06-04
Sorry, @eregon can you edit the change log to also credit @lemire?