@lemire 2021-06-04
Just in case this needs saying... the license on my blog is "public domain"... meaning that taking from the source code of my blog and doing "whatever" is encouraged.