@eregon 2021-06-04
Done, here is what I have (should merge soon): https://github.com/oracle/truffleruby/compare/master...eregon:faster-integer-length